requirements.txt
# Core Deep Learning
torch
torchvision
torchaudio

# Backbone Model Library
timm

# Data & Image Processing
numpy
opencv-python-headless
pillow
kornia

# Metrics & Visualization
matplotlib
seaborn
scikit-learn

# Utilities
tqdm